+++ packages/api-headless-cms-ddb/src/operations/model/index.ts
@@ -19,6 +19,11 @@ interface PartitionKeysParams {
 }
 const createPartitionKey = (params: PartitionKeysParams): string => {
     const { tenant, locale } = params;
+    if (!tenant) {
+        throw new WebinyError(`Missing tenant variable when creating model partitionKey.`);
+    } else if (!locale) {
+        throw new WebinyError(`Missing locale variable when creating model partitionKey.`);
+    }
     return `T#${tenant}#L#${locale}#CMS#CM`;
 };
 
@@ -85,7 +90,7 @@ export const createModelsStorageOperations = (params: Params): CmsModelStorageOp
             return model;
         } catch (ex) {
             throw new WebinyError(
-                ex.messatge || "Could not update model.",
+                ex.message || "Could not update model.",
                 ex.code || "MODEL_UPDATE_ERROR",
                 {
                     error: ex,
@@ -106,7 +111,7 @@ export const createModelsStorageOperations = (params: Params): CmsModelStorageOp
             return model;
         } catch (ex) {
             throw new WebinyError(
-                ex.messatge || "Could not delete model.",
+                ex.message || "Could not delete model.",
                 ex.code || "MODEL_DELETE_ERROR",
                 {
                     error: ex,
@@ -128,7 +133,7 @@ export const createModelsStorageOperations = (params: Params): CmsModelStorageOp
             return cleanupItem(entity, item);
         } catch (ex) {
             throw new WebinyError(
-                ex.messatge || "Could not get model.",
+                ex.message || "Could not get model.",
                 ex.code || "MODEL_GET_ERROR",
                 {
                     error: ex,
@@ -153,7 +158,7 @@ export const createModelsStorageOperations = (params: Params): CmsModelStorageOp
             return cleanupItems(entity, items);
         } catch (ex) {
             throw new WebinyError(
-                ex.messatge || "Could not list models.",
+                ex.message || "Could not list models.",
                 ex.code || "MODEL_LIST_ERROR",
                 {
                     error: ex,