+++ packages/api-i18n-ddb/src/definitions/systemEntity.ts
@@ -19,6 +19,9 @@ export default (params: { context: I18NContext; table: Table }): Entity<any> => 
             version: {
                 type: "string"
             },
+            tenant: {
+                type: "string"
+            },
             ...attributes
         }
     });