+++ packages/api-page-builder-import-export/src/graphql/crud/pageImportExportTasks.crud.ts
@@ -59,7 +59,7 @@ export default ({ storageOperations }: PageImportExportPluginsParams) =>
                 });
 
                 const tenant = context.tenancy.getCurrentTenant();
-                const locale = context.i18nContent.getLocale();
+                const locale = context.i18nContent.getCurrentLocale();
 
                 const params = {
                     where: {
@@ -99,7 +99,7 @@ export default ({ storageOperations }: PageImportExportPluginsParams) =>
                 });
 
                 const tenant = context.tenancy.getCurrentTenant();
-                const locale = context.i18nContent.getLocale();
+                const locale = context.i18nContent.getCurrentLocale();
 
                 const { sort, limit } = params || {};
 
@@ -146,7 +146,7 @@ export default ({ storageOperations }: PageImportExportPluginsParams) =>
                 const pageImportExportTask: PageImportExportTask = {
                     ...data,
                     tenant: context.tenancy.getCurrentTenant().id,
-                    locale: context.i18nContent.getLocale().code,
+                    locale: context.i18nContent.getCurrentLocale().code,
                     id,
                     createdOn: new Date().toISOString(),
                     createdBy: {
@@ -287,7 +287,7 @@ export default ({ storageOperations }: PageImportExportPluginsParams) =>
                 const pageImportExportSubTask: PageImportExportTask = {
                     ...data,
                     tenant: context.tenancy.getCurrentTenant().id,
-                    locale: context.i18nContent.getLocale().code,
+                    locale: context.i18nContent.getCurrentLocale().code,
                     id: id,
                     parent: parent,
                     createdOn: new Date().toISOString(),
@@ -366,7 +366,7 @@ export default ({ storageOperations }: PageImportExportPluginsParams) =>
                     rwd: "r"
                 });
                 const tenant = context.tenancy.getCurrentTenant();
-                const locale = context.i18nContent.getLocale();
+                const locale = context.i18nContent.getCurrentLocale();
 
                 const params = {
                     where: {
@@ -407,7 +407,7 @@ export default ({ storageOperations }: PageImportExportPluginsParams) =>
                 });
 
                 const tenant = context.tenancy.getCurrentTenant();
-                const locale = context.i18nContent.getLocale();
+                const locale = context.i18nContent.getCurrentLocale();
 
                 const listParams: PageImportExportTaskStorageOperationsListSubTaskParams = {
                     where: {